1) changed from GList to GSlist for hblisteners.  Have no need for doubly linked lists 2) added code to notice that a node is dead - and cause it to be activated -it works 3) fixed a not completely obvious bug in unref/freeing of HbListener objects
So, this is the first version that the test code actually listens for missing heartbeats, and notices that they are missing.
It still doesn't send any (which explains why they're missing) - but this is good progress.  The reception of packets and dispatching them according to their type is more complex than sending heartbeats.  So, I'm reasonably happy :-D