Made 'os' discovery happen first _and unsolicited_.
This is because we really need the OS info before almost anything.
In particular we need it in the best practice rules to form the
URL of the details on the IT Best Practices project web site.

Before, when it returns is subject to the vagaries of the scheduler.
It's not *guaranteed* to be return its result first, but it's cheap
and it's definitely always started before any other discovery scripts.