Updated builds and magic installer
All the builds work and everything installs except for the CMA on Fedora.

There are likely still issues about the version of Neo4j being installed via RPMs.
If I installed via tar ball, then that wouldn't be an issue and the CMA could
be installed on Fedora. But RPMs are the right answer - if Neo4j keeps up
with their part. See Neo4j#5388.