Updated my personal "checkandput" command
I use it to validate my source before pushing upstream.
It does this by running it through clang, checking the python with pylint
making sure the headers all compile by themselves, and that it
passes our test suite.

It pushes changes in the current repository upstream and also updates the web site
if the above tests all pass.