Beginning process of creating 1.0.3 release
And by updating these now, I can use some not-yet-committed code to build
and install pre-releases in a way that's less public.

Before I had the problem with at least one person installing the "release"
before it was ready. This should remedy that. You can install the
pre-release, but not accidentally...