Changing MAC/IP relationship causes an exception.
Details are in https://trello.com/c/ySlAkWJC
s appears to be caused by a bug which was documented in the code this way:

Although this bug didn't seem obvious to me at the time, it does now.
Optional parameters were passed in the wrong order without being
explicitly named. That caused the relationship type to be confused
with the far-end object of the relationship.
Since the relationship type is a string, that would cause this error.
This appears to have fixed it.