Added instance to LLDP/CDP JSON
It had been added to the nanoprobe for the other cases, but this
data comes through a different route, so it had to be added to the
return structure from the nanoprobe and also to the CMA.

While I was at it, I fixed a bug regarding LLDP port and chassis id
caching. It was not caching properly because it wasn't retrieving
the port and chassis id correctly.