Dropped Libsodium version constraint
It's not an API question - it's a question of a bug that I'd rather not eventually
be the victim of. The most probable effect of this bug is that it causes valgrind
to notice it (but I turned that particular warning off).

But if it does cause a problem, that problem will be a crash. It all depends on the
page boundaries work out. If a message ends exactly at a page boundary, and the nex
page in the address space is missing (isn't allocated), then the app will crash.

Not likely, but certainly possible...

But openSUSE 13.2 only supplies version 1.0.