Stopped using prettyprinting for discovery diffs
Turns out that prettyprinting formatting has changed over the
releases, so it would fail on some old versions of jsonlint.

It now compares without any spaces, then if there's a difference
pretty prints and diffs the prettyprinted result to show
what's broken. That way we're comparing like to like as far
as versions are concerned and in addition, you can actually
read the output, which you couldn't do on the version with
no spaces at all.