Issue #19: Introduce UNWARN event and add event sending to best practices
This commit introduces a new event type called an UNWARN which will signal that
a best practices failure has been resolved. It also reorders the numbers for
the events so WARN and UNWARN are close together. The UNWARN events will be
send when when a failed best practice is resolved to the pass, ignore, or NA
state. Likewise the event sending code will send WARN events when a best
practices failure is generated after the rule was not previously detected
(none) or  detected as a pass, ignore, or NA.