Use Neo4j Auth: https://trello.com/c/NGuuSM8D
We need to use Neo4j Authentication. To do this we create a random
password for 'neo4j' when we first start up, and cache it on disk
(in clear text) for future use.

This is OK since the file and the directory it's in are only readable
by 'root' or 'assimilation'.

You can change the password with the assimcli command. Don't use
the underlying mechanisms or we won't have a copy of it to use
when we need to connect to Neo4j.