Add debugging for Travis.ci problem w/Neo4j
It appears that /var/lib/neo4j/conf/neo4j-server.properties is somehow being cached.
That's the most probable cause for the 404 seen here: https://travis-ci.org/assimilation/assimilation-official/builds/101471798
This change looks for dbms.security settings to see if they disable server security or enable it.
We used to disable it, but now I've fixed it so it *requires* it. Neo4j version seems right, and the earlier 'cat' of the server.properties file shows it as correct. But it's now behaving as though it's set to false.