Fixed ConfigContext floating point output for ints
https://trello.com/c/8ix8ZghW

If a floating point number is exactly an integer, then the %g format
won't put any decimal points in the output. This means that when
we read it back in, we'll think it's an int type object.

So, the code now makes sure a decimal point appears in floating
point values. So far, this just affects some of our test code
related to best practice scores.

I don't think we have any other examples of floating point numbers
in our code.