Support creating LLDP and CDP test code
This is a new python class and underlying C code for opening capture
files, reading packets and so on.

There is no code yet to test the LLDP or CDP code using capture
data, but this makes it relatively easy to write in the future.
I just wanted to get this into source control before I lost it ;-).