Added some LLDP/CDP processing tests
Really ought to do the same thing I do elsewhere in some of the
JSON-based tests - compare the two after reformatting, or at least
reformat them and print them if different before raising/asserting
the error.

But this is perfectly functional. Just could be easier to diagnose any
future problems or changes if it were formatted prettier.