fixed long-standing nanoprobe crash-on-shutdown?
My best guess is that a timer popped after we got a signal but before we
started shutting down.

What I did in the code was NULLed out a timer on the callback code - so
that the callback-on-shutdown won't happen.

I tried for a few hundred times to make this happen after putting in
this change and could not make it happen. It might be fixed, and it
might not.

Since the ARP code and the LLDP/CDP code are very similar, and the LLDP
code doesn't do this, my guess was that this was in the differences.
What's different between the two types of code was that the ARP code has
a timer for returning results. The LLDP/CDP code does not. So my guess
was that this was likely in the timer code.