Added a rule for recognizing Oracle for monitoring
Not sure if it's right, but it's a start.

If it isn't right, it may be messier to discover if Oracle is running.

My concern is that ora_pmon_foo may not be listening on the network.
If it's not, then this rule won't work, and it may get complicated...