Added gem,npm,php,pip to "packages" discovery
That means we now collect package info for these package types
	deb - Debian
	gem - Ruby GEMs
	npm - Node.js package manager
	php - PHP packages
	pip - PIP Python package installer
	rpm - RPM Package Manager (recursive acronym)
              (formerly the Redhat Package Manager)

Deferring Perl until a later time - it is 100x as expensive as all of
the above put together.

Java doesn't have a central package manager. So, we'll have to get the
CLASSPATHs of the running Java programs to get version information for
these Java programs. This means it can't be run until after we know what
Java services we're offering.