JSON floats must end in .0 for integral values
In other words, when a floating point number has a value which has no
fractional part (is an integer), then JSON requires that it end in ".0".

In C, the "%g" format never adds a decimal place to such a value, and we
had added a decimal to the end "1.", but not a zero. Now the code adds
".0" to the end so that strict JSON requirements are satisfied.

This has been broken all along, but recently discovered because
we added floating point values to the config file.