NICNode has no attribute 'ifname' bbcXJhhQ
https://trello.com/c/bbcXJhhQ

This was an instance of a more general problem. If we tried to construct
a node with more information than the node already in the database had
in it (that is, more fields), then these other fields would never get
set.

So here's how to make it happen:

	Construct an object without some optional attributes
	Create the same object with some optional attributes
		not supplied when it was first created

This (supposedly-updated) object will not have these optional attributes set.

If you set them normally, then they'll get set. But if you try and
construct it using load_or_create, then this will happen.