Making docker discovery work!
We now discover Docker instances and then trigger discovery for the
containers, and create them in the database almost exactly like they
were "real" systems.

This code should support Vagrant just as well, _provided_ that the
vagrant nodes were created by root - but I haven't checked for
that yet. This is something I'll eventually be able to work around, but
for now, it's a limitation.