Forgot to add vagrant discovery agent. OOPS!
It now captures the user and group id of the Vagrantfile.

If the userid which created the Vagrantfile is the same userid as
created the VM, then we will eventually be able to do things with
vagrant just like we can with Docker now.

This will require that we add the capability for the nanoprobe to
run a command as a given user. We ought to have that capability anyway
so that commands that don't need root permissions don't have them.