Discovery always invoked with 'discoverychanged'
We changed the discovery API so that we now pass a 'discoverychanged'
parameter to each discovery module _and_ we _always_ call it whether
anything has changed or not.

Discovery modules now have access to the old value for this particular
discovery parameter because the database is updated _after_ discovery
processing is complete. It's part of the same transaction, but done
after calling the individual discovery modules.

This means I should be able to get rid of the old saved attributes for
checksums (at least I think it's checksums). But this will do for now.