allow Neo4j alternate authentication for testing
Normally, we use the local cache of authentication information for
connecting to Neo4j. However, for testing, we need to give some
different authentication because the CMA is on a different machine than
the tests are being run on, and we need to do queries against _that_
instance of Neo4j, not the local instance.

So I added some parameters to allow the appropriate objects to be given
parameters to specify authenticaiton information rather than letting it
default to the cache files on the local machine (which you may not even
be able to read when testing).