No longer explicitly keep OLD_checksums value.
We now update the checksums value at the end of the transaction, so the
old value is available to us.

This is the right way to solve these kinds of one-generation problems.

The right way to solve the bigger problem is to change the schema to
keep the old values forever, and link them from the newest
values - creating a linked list of values over time.