tcpdiscovery missing from graph: trello.com/c/nm3IYSsk
Bill Welch discovered a bug where some discovery gets translated into a
pyNetAddr because it's an IP address. This particular discovery was in
an argument list to a process. This meant not only was it an
unacceptable value for Neo4j (a pyNetAddr), but it made the elements of
the array to be of mixed types - which is also unacceptable to Neo4j.

Although the transformation into an IP address was questionable, fixing
that might have other consequences which are potentially far-reaching.

Instead, the approach in this patch was to make sure everything was of
an acceptable type to Neo4j when we persisted them to Neo4j. In this
case, that meant transforming it to a string. This means it will come
back as a string when put into a simple attribute or an array element,
but when the JSON is revived into structures, that will have the type
pyNetAddr. I don't know when/how this will be a problem, but it's a
potential future problem.

The patch appears to work for Bill. It will also fix other occurances of
type mismatches when the type is pyNetAddr. This probably means some of
the other places where this has been worked around could be removed and
let this mechanism "solve" the problem.