More Store progress
I've started the process of changing most of the updates to use the
Neo4j transactions the way they were intended to be used.
The exception is modification of attributes. Those are batched to be
performed at the end.

Probably should make sure that these updates don't get overwritten
somehow by queries bringing in the same objects that I already have.