The CMA tests all work w/new Ring architecture(!)
I haven't re-run the Store tests - so I could have broken something. But
the CMA tests rely heavily on the Store for correct behavior - so it
can't be _too_ badly broken. I just wanted to push all this stuff to the
repository so that I couldn't lose it.

This new architecture changes it so that there is no graph relationship
between a ring and it's members. That relationship is maintained via a
label. Each ring has a unique label which identifies it. They all follow
the pattern of with "Ring_${ringname}" - where $ringname is the unique
name given to the ring. This means we need to give the rings "nice"
names suitable for giving to a label name.

At the moment, there's only the one ring, so that's not an issue...

In addition, each Drone->Drone relationship has the same name now, with
a relationship attribute of 'ring_name' distinguishing membership in one ring versus
another in the relationships in the database.

Although the idea of rings and how they work is unchanged, the
underlying implementation is now quite different.