Add alpine discovery to 'packages' discovery
It's sort-of tested. But since Alpine is one of the favored
environments for Docker, it seems good to include Alpine package
discovery to it.

By sort-of tested I mean that the main 'sed' format mangling
has been tested against a list of alpine packages.
But the whole thing hasn't been tested against an alpine Docker
image.