Added code to persist JSON directly to filesystem
No one uses it for anything, but it does seem to work.

The abstract base class should persist using any mechanism.
The only concrete implementation of this class is one that persists all
the data into flat files in a UNIX-like filesystem.

It uses syncfs() and/or fsync() to make sure the bits wind up on disk
for real.

It has lots of lovely Python wrappers to make sure it all is pretty and
natural from the Python perspective.

There is sort of a search facility, but it doesn't really do anything
yet.

Need to add some tests for it to our test suite.