Improved documentation and spacing for the FSA clientlib/fsprotocol.c
This is to make the FSA clearer. Everything was literal white space changes including removing tabs from the layout of the FSAs except for using the CLOSEnNOTIME macro in a few places where it replaced exactly what it was intended to replace.

No semantic changes were intended. Hopefully none occurred...