ci(travis): build ci/cd pipeline
Dockerfiles stored in ci/Dockerfiles/ define the environment that must
be set up prior to building the source. The source code is mounted into
the running docker container and using cmake, the source is built. cpack
is executed afterwards to generate packages which should be moved to
/build_artifacts. Since the running container mounts
(/tmp/ba:/build_artifacts), these packages can be also be shared into a
new stock-OS docker container where they are installed and tested.